Added yelpSearch to helper functions file to clean up widgets.js
